I am working on fixing the `withOpacity` deprecation and refactoring the registration screens.
